Fix voldemort types in public interface
The `Fds` struct used in `ListenFds` in the `pingora::services::Service` trait
is not nameable. See https://rust-lang.github.io/rfcs/2145-type-privacy.html#~~~-voldemort-types-its-reachable-but-i-cant-name-it

This change makes both `ListenFds` and `Fds` public.

As a side-effect all of the methods on `Fds` are now public as well but are not documented.

Fixes: https://github.com/cloudflare/pingora/issues/145

Replicated-from: https://github.com/cloudflare/pingora/pull/146
Includes-commit: 6f835a357b448c35362cf2197a1160c72505d2dd