add builder api for pingora listeners:
This change takes the existing API for ListenerEndpoint and modifies it to use a typical idiomatic builder style.

There are additional improvements which I feel can be made here, but I believe that it would be better to do them in subsequent commits in order to keep changes as incremental as possible.

In particular, I think that making socket options into sub-builder is probably a good move.