feat(cache): Allow using in-memory compression dicts
This PR allows setting the compression dictionary used for headers in
`pingora-cache` from in-memory bytes instead of an external file.

This is useful if the dict is already embedded in the binary, or is
retrieved from remote storage dynamically.

Also cleans up the code a bit by making the constant private, moving the
initialisers to the same file and re-exporting them to keep the logic
contained, and by slightly simplifying the file loading code.