feat(core): Log runtime names during Server shutdown
Log the names of runtimes during Server shutdown.

Motivation: server shutdown can take quite a long time if the runtimes
don't exit.
It's hard to debug why this is the case, because there is no indiciation
in the logs which service is pending.

This small change modifies the server to log the name of the service
whose runtime is being waited for.