Signal that the response is done when body write finishes
Since this change response_duplex(_vec) will return true when the body
write finishes even if the end_stream flag is not set. This change makes
pingora-proxy stop polling from the downstream connection when the
response finishes already.

Before this change, in a rare case where the h2 upstream responds the
complete body according to its content-length header but end_stream is not
received yet, the h1 downstream would already consider sending the next
request over the reused connection because h1 doesn't rely on any explicit
end_stream signal when content-length is used. In this case, the next
request would cause an error.