Require WritePermit to release lock, fix dangling detection
Dangling cache locks could occur when there were simply many readers at
once waiting for the lock, the reason being the available permits could
immediately be used up upon unlocking and before the WritePermit is
dropped, resulting in a false positive dangling lock.

Instead, just mark the WritePermit as finished when the cache key lock
is unlocked as a result of having the permit. The most intuitive way to
do this is to "turn in" the WritePermit back to the CacheKeyLock that
created it when unlocking, so that the CacheKeyLock can mark it as
finished.