Allow reusing session on errors prior to proxy upstream
Allow session reuse if configured in `fail_to_proxy`. Applies only to
errors prior to proxying upstream, for now.

Previously any short-circuiting error during request processing would
result in closing the connection to downstream. This is unnecessary if
there was an internal error that simply resulted in an error response
being written downstream. Note that any `set_keepalive` on the Session
itself is still respected, and this API currently does not overwrite
reuse decisions made while proxying upstream.

The user is responsible for implications of reusing the session in this
manner, such as draining any remaining request body.