proxy_upstream_filter tries to reuse downstream
Currently a non-erroring proxy_upstream_filter result will not try to
reuse the downstream session. Change the default to try to reuse the
session; if the session should not be reused, the user implemented
filter can disable keepalive.

Also adds a `drain_request_body` util function to the server Session to
make it more convenient for the caller to make the session reusable.