Fixes HTTP1 client reads to properly timeout
The initial read(s) for reading in the HTTP request headers
from the client did not make use of the `read_timeout`, and
a client could cause a connection to hang forever.

Also, there was no way to set the `read_timeout` for an
`HttpSession` prior to this initial read, despite there
being a `set_read_timeout`, so this commit also sets a
"sensible default" of 60 seconds, which matches NGINX's
default value for `client_header_timeout` and `client_body_timeout`.

Includes-commit: 0d48675de346283ba23d41abb18569b13d6e3386
Replicated-from: https://github.com/cloudflare/pingora/pull/539