Allow modules to receive HttpTask::Done, flush response compression
This change adds a `response_done_filter` hook for downstream HTTP modules
to handle the HttpTask::Done type, which may be sent in lieu of a body +
end flag, for example in certain caching request paths.

In the `ResponseCompression` module this could manifest as un-flushed
body bytes. This change duplicates the `response_filter` logic currently
used for upstream response compression.