Allow spawning background subrequests from main session
The proxy Session now acquires a SubrequestSpawner from the HttpProxy
app logic. This allows spawning subrequests using the original Session.

Whether the spawner is assigned to the session is dependent on an
allow_spawning_subrequest filter, as few sessions may actually need to
make use of such a spawner.

Also allows passing an arbitrary UserCtx struct to the subrequest for
users to pass specialized data from the parent to the subrequest. The
existing subrequest spawning API is revamped with a Ctx::builder for
subrequest configurability in the future.

Subrequest spawning should be considered pre-alpha and experimental, and
APIs are highly volatile.