Fix chunked trailer end parsing
httparse itself does not treat the terminating chunk (0 + CRLF) any
differently when parsing its chunk size, and the body reader does not
validate the following CRLF required to close the body.

Additionally, that current parsing scheme will not consider trailers
before the end CRLF. Now the trailers are considered but simply
discarded. The CRLFs between the trailers (and now, the CRLFs after the
body payloads) are validated, however.