Fix CacheLock status timeout conditions
Introducing a wait timeout meant readers could replace a write lock that
was timed out. However, if the writer subsequently errored or finished,
then that lock status would be replaced in CacheLock, causing readers to
busy wait and be unable to replace that lock. This change prevents
overriding the Timeout status in CacheLock if it was set.

Additionally, this change fixes an issue where an already expired lock
may not have the Timeout status applied by subsequent readers.