Mark previously too large chunked assets as cacheable
Too large assets without content-length could previously be skipped as
uncacheable indefinitely by the cacheable predictor. This change moves
the max file size tracking outside of the miss handler itself such that
it can also be used to track body bytes after caching was disabled.

This also adjusts the cache inner structs such that cache key and the
cacheable predictor are accessible, so that an asset can be remembered
as cacheable even if caching was disabled.

Additional safeguards are also in place to ensure that the NeverEnabled
reason is not overridden when disabling cache, as it does not make sense
to disable cache when it was never enabled in the first place.