Subrequest sessions, clear body headers if no input
Subrequests are now a separate server Session type, instantiated with
corresponding SubrequestHandles that can send or receive HttpTasks to
or from the subrequest. This makes it possible to communicate with
created subrequests.

Additionally the subrequest ctx now has a BodyMode input. If unset the
background subrequest will clear request headers related to the request
body to prevent issues where the upstream might expect body.