feat: Virtual L4 Streams
Introduce a virtual stream abstraction that allows developers to easily
integrate custom network implementations or protocols into the L4 connector layer
used by pingora-proxy.

This is necessary because while the API allows for custom connector
implementations, these at the moment must resolve to an actual TcpStream
or a unix domain socket. Before this change it was not possible to have streams
backed by custom implementations.

- Changes:

The `RawStream` enum is extended with an additional `Virtual` variant,
which holds a `VirtualSocketStream`.

`VirtualSocketStream` is a wrapper around any type implementing the new
`VirtualSocket` trait, which mostly is a simple wrapper trait around
`AsyncRead` and `AsyncWrite`, plus the ability to set socket options.

- Performance implications:

Since `RawStream` already is an enum with multiple variants on Unix, due to the domain sockets, adding an additional variant should not impact performance in any notable way.

- Example:

A new example at pingora-proxy/examples/virtual_l4.rs demonstrates basic
usage of the functionality.