Fix cache admission on header only misses
Previously if the downstream errored or went away during header only
misses, this would fail cache admission of the upstream pull. Now a
HeaderOnlyMiss and DoneMiss phase are provided, which continue
allowing the upstream cache miss body to be admitted into cache,
while preventing incorrectly passing those body tasks to the
downstream session where (in the case of the h2 session) they would be
improperly proxied.