Restore daemonize STDOUT/STDERR when error log file is not specified
Pingora documentation indicates that when the error_log is not specified, error logs are sent to STDERR (https://github.com/cloudflare/pingora/blob/main/docs/user_guide/conf.md?plain=1#L21).

However, when pingora is being ran as a daemon, error messages are being suppressed, and never make it to the error log, because the daemonize crate redirects all of the streams to /dev/null (https://github.com/knsd/daemonize/blob/master/daemonize/src/lib.rs#L270-L272).

Restoring this behaviour seems to be as simple as setting the stderr field of the daemonize struct to the daemonize.Stdio Keep implementation (which basically doesn't redirect any of the standard streams).

In this commit we're also restoring STDOUT, since it seems more logical for the caller to be able to drop that stream, than not being able to access it at all in the first place (which might be useful for debugging use-cases).